auth: done login and filter, remain authorization

+++ src/main/java/com/codese/SpringBootBackendBoilderplate/helper/jdbc_mapper/UserMapper.java
@@ -10,7 +10,7 @@ public class UserMapper implements RowMapper<User> {
     @Override
     public User mapRow(ResultSet resultSet, int i) throws SQLException {
         User user = new User();
-        user.setId(resultSet.getInt("id"));
+        user.setId(resultSet.getLong("id"));
         user.setUsername(resultSet.getString("username"));
         user.setPassword(resultSet.getString("password"));
         return user;
